Removed secret file, moved accounts to URL
