+++ commands/auth/axios-authenticated.js
@@ -1,0 +1,7 @@
+const axios = require('axios')
+
+module.exports = config => axios.create({
+  headers: {
+    Authentication: `${config.key}:${config.secret}`
+  }
+})
