--- commands/keys/axios-authenticated.js
@@ -1,7 +1,0 @@
-const axios = require('axios')
-
-module.exports = keys => axios.create({
-  headers: {
-    Authentication: `${keys.key}:${keys.secret}`
-  }
-})
