+++ commands/machines/machines.js
@@ -2,6 +2,6 @@ const url = require('url')
 
 module.exports = ({get, config}) => ({
   getMachines: () =>
-    get(url.resolve(config.endpoint, '/api/machines'))
+    get(url.resolve(config.endpoint, `/api/accounts/${config.account}/machines`))
       .then(response => response.data)
 })
