+++ commands/tasks/tasks.js
@@ -1,22 +1,25 @@
 const url = require('url')
 const querystring = require('querystring')
 
-module.exports = ({get, put, config}) => ({
-  getTasks: machine =>
-    get(url.resolve(config.endpoint, '/api/tasks'), querystring.stringify({machine: machine && machine.id}))
-      .then(response => response.data),
+module.exports = ({get, put, config}) => {
+  const address = url.resolve(config.endpoint, `/api/accounts/${config.account}/tasks`)
 
-  runTask: ({machine, command, output, taskName, tier}) => {
-    const body = {
-      command,
-      output,
-      machine,
-      taskName,
-      tier
+  return {
+    getTasks: machine =>
+      get(address, querystring.stringify({machine: machine && machine.id}))
+        .then(response => response.data),
+
+    runTask: ({machine, command, output, taskName, tier}) => {
+      const body = {
+        command,
+        output,
+        machine,
+        taskName,
+        tier
+      }
+
+      return put(address, body)
+        .then(response => 'Response: Task queued successfully')
     }
-
-    let address = url.resolve(config.endpoint, '/api/tasks')
-    return put(address, body)
-      .then(response => 'Response: Task queued successfully')
   }
-})
+}
