+++ eureka.config.js
@@ -1,4 +1,8 @@
 module.exports = {
+  key: 'Key',
+  secret: 'Secret',
+  account: 'eb2e695df97f4b5e83c1074261d3bb3e',
+
   // DEV:
   endpoint: 'http://eureka-dev.us-west-2.elasticbeanstalk.com'
