--- eureka.credentials.json
@@ -1,4 +1,0 @@
-{
-  "key": "Key",
-  "secret": "Secret"
-}
